Add m_size field to the mbuf, initialised in mb_ctor_clust(), mb_ctor_pack(), m_pkthdr_init(), m_extadd(), mb_dupcl(), m_move_pkthdr(), m_dup_pkthdr(), sendfile(), m_extaddref(), m_init(), and m_cljset().
Add M_ASSERTSIZE() which validates that m_size has a reasonable sensible
set of contents -- called in m_free().

Having discovered that m_size was frequently MHLEN rather than MLEN,
sweep the tree to replace manual frobbing of M_PKTHDR with either an
assertion that it is set (rather than setting it), or m_pkthdr_clear() to
safely remove it, clearing tags and updating m_size.